# onidelcloud/onidel-api-docs.yaml
openapi: 3.0.3
info:
  title: Onidel Cloud API - OpenAPI 3.0
  description: |-
    Welcome to the **Onidel Cloud API** documentation. This API allows developers to seamlessly integrate with the Onidel Cloud platform, enabling them to manage cloud resources efficiently and securely. With our API, you can automate processes, retrieve detailed account information, and manage virtual machines (VPS), networking, storage, and more.
  termsOfService: https://docs.onidel.com/policies/terms-of-service
  contact:
    email: support@onidel.com
  version: 1.2.0
servers:
  - url: https://api.cloud.onidel.com
...
paths:
...
  /ssh_keys:
    get:
      tags:
        - SSH Key
      summary: List SSH Keys
      description: Retrieve all SSH keys for a team
      operationId: getSSHKeys
      parameters:
        - name: team_id
          in: query
          description: Team ID to filter SSH keys. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHKeys'
        '401':
          description: Unauthorized
        '404':
          description: Not found
      security:
        - api_key: []
    post:
      tags:
        - SSH Key
      summary: Create a new SSH Key
      description: Add a new SSH key to the team
      operationId: createSSHKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewSSHKey'
      responses:
        '201':
          description: SSH key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHKeyResponse'
        '400':
          description: Bad request - missing or invalid parameters
        '401':
          description: Unauthorized
      security:
        - api_key: []
...
  /object-storage:
    get:
      tags:
        - Object Storage
      summary: List Object Storage Services
      description: Retrieve all object storage services for the team
      operationId: getObjectStorageServices
      parameters:
        - name: team_id
          in: query
          description: Team ID. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: '#/components/schemas/ObjectStorageService'
        '401':
          description: Unauthorized
      security:
        - api_key: []
...
  /object-storage/{service_id}/buckets:
    post:
      tags:
        - Object Storage
      summary: Create a Bucket
      description: Create a new bucket in the object storage service
      operationId: createBucket
      parameters:
        - name: service_id
          in: path
          description: Object storage service UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - bucket_name
              properties:
                team_id:
                  type: string
                  format: uuid
                  description: Team ID. Default team will be used if not provided.
                  example: 96616865-475e-4191-aba5-f89f0a4b71fe
                bucket_name:
                  type: string
                  description: Name of the bucket to create
                  example: my-bucket
                versioning:
                  type: boolean
                  description: Enable versioning on the bucket
                  default: false
                object_lock:
                  type: boolean
                  description: Enable object lock on the bucket
                  default: false
      responses:
        '201':
          description: Bucket created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    description: Access keys for the newly created bucket
                    items:
                      type: object
                      properties:
                        access_key:
                          type: string
                        secret_key:
                          type: string
        '400':
          description: Bad request (invalid name, bucket already exists, too many buckets)
        '401':
          description: Unauthorized
        '404':
          description: Service not found
      security:
        - api_key: []
...
  /network/reserved_ips:
    get:
      summary: List reserved IPs
      description: Get all reserved IPs for the authenticated user's team or specified team
      tags:
        - Reserved IPs
      security:
        - api_key: []
      parameters:
        - name: team_id
          in: query
          description: Team ID to fetch reserved IPs for. If not provided, uses the user's primary team
          required: false
          schema:
            type: string
            format: uuid
        - name: location
          in: query
          description: Filter reserved IPs by location name (case-insensitive)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of reserved IPs
          content:
            application/json:
              schema:
                type: object
                properties:
                  reserved_ips:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReservedIP'
        '401':
          description: Unauthorized - user doesn't have read permission for the team
    post:
      summary: Create a new reserved IP
      description: Allocate a new reserved IP address in the specified location
      tags:
        - Reserved IPs
      security:
        - api_key: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                team_id:
                  type: string
                  description: Team ID to create the reserved IP for. If not provided, uses the user's primary team
                ip_type:
                  type: string
                  enum:
                    - ipv4
                    - ipv6
                  description: IPv4 (/32) or IPv6 (/64) address
                location:
                  type: string
                  description: Location name where the IP should be allocated
                name:
                  type: string
                  maxLength: 64
                  description: Custom name for the reserved IP (defaults to IP address if not provided)
              required:
                - location
      responses:
        '200':
          description: Reserved IP successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  rip_id:
                    type: string
                    format: uuid
                    description: ID of the newly created reserved IP
                  ip_addr:
                    type: string
                    description: Newly created reserved IP address
                required:
                  - rip_id
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: string
                    enum:
                      - UNSUPPORTED_LOCATION
                      - IP_NAME_TOO_LONG
                      - NOT_SUPPORTED
                    description: Error code indicating the specific validation failure
        '401':
          description: Unauthorized - user doesn't have write permission for the team
        '402':
          description: Payment required - insufficient balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: string
                    enum: [INSUFFICIENT_BALANCE]
        '403':
          description: Forbidden - quota exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: string
                    enum: [RESERVED_IP_QUOTA_EXCEEDED]
        '404':
          description: Resource not available
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: string
                    enum: [RESOURCE_NOT_AVAILABLE]
                    description: No available IP addresses in the specified location
...
components:
  schemas:
    ObjectStorageService:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        location:
          type: string
        region:
          type: string
        endpoint:
          type: string
        capacity:
          type: integer
          description: Storage capacity in KB
        used_capacity:
          type: integer
          description: Used capacity in KB
        status:
          type: string
          enum:
            - active
            - suspended
        created_at:
          type: string
          format: date-time
        renewal_date:
          type: string
          format: date-time
...
    SSHKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: a3f2e5d8-1234-4abc-9876-543210fedcba
        created:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        name:
          type: string
          example: "Production Server Key"
        ssh_key:
          type: string
          example: "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC... user@hostname"
      required:
        - id
        - created
        - name
        - ssh_key
    SSHKeys:
      type: object
      properties:
        ssh_keys:
          type: array
          items:
            $ref: '#/components/schemas/SSHKey'
    SSHKeyResponse:
      type: object
      properties:
        ssh_key:
          $ref: '#/components/schemas/SSHKey'
    NewSSHKey:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
          example: c1d45377-b2a9-4407-ab8d-6909c34dfaac
        name:
          type: string
          example: "Production Server Key"
        ssh_key:
          type: string
          example: "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQC... user@hostname"
      required:
        - team_id
        - name
        - ssh_key
...
    ReservedIP:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Reserved IP UUID
        name:
          type: string
          description: Name of the reserved IP
        location:
          type: string
          description: Location name where the IP is allocated
        status:
          type: string
          enum: [active, suspended]
          description: Current status of the reserved IP
        attachment:
          type: object
          nullable: true
          description: VM attachment details, null if not attached
          properties:
            id:
              type: string
              format: uuid
              description: UUID of the attached VM
            name:
              type: string
              description: Name of the attached VM
        billing_cycle:
          type: integer
          description: Billing cycle in months
        renewal_date:
          type: string
          format: date-time
          description: Next renewal date
        last_renewal:
          type: string
          format: date-time
          description: Last renewal date
        recurring_amount:
          type: number
          format: decimal
          description: Recurring billing amount
        currency:
          type: string
          description: Currency code for billing
        total_billed:
          type: number
          format: decimal
          description: Total amount billed
        ip_addr:
          type: string
          format: ipv4
          description: The reserved IP address
        suspension_reason:
          type: string
          description: Reason for suspension (empty string if not suspended)
      required:
        - id
        - name
        - location
        - status
        - billing_cycle
        - renewal_date
        - last_renewal
        - recurring_amount
        - currency
        - total_billed
        - ip_addr
        - suspension_reason
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer